+++ src/components/common/SelectBox.tsx
@@ -1,11 +1,9 @@
 import { MouseEvent, useState } from 'react';
-import { createPortal } from 'react-dom';
 import { UseFormSetValue, UseFormTrigger } from 'react-hook-form';
 
 import Icon from './Icons';
 interface PropsType {
   items: string[];
-  containerRef: React.RefObject<HTMLDivElement>;
   currentItem: string;
   changeItem: (item: string) => void;
   size: 'large' | 'small';
@@ -16,7 +14,6 @@ interface PropsType {
 
 const SelectBox = ({
   items,
-  containerRef,
   currentItem,
   changeItem,
   size,
@@ -49,7 +46,7 @@ const SelectBox = ({
   };
 
   return (
-    <div>
+    <>
       <div
         onBlur={onBlur}
         tabIndex={4}
@@ -66,31 +63,27 @@ const SelectBox = ({
           {!isOpen && <Icon.Down size={size === 'large' ? 'large' : 'small'} />}
         </span>
       </div>
-      {containerRef.current &&
-        createPortal(
-          <div
-            className={`${isOpen ? 'visible' : 'invisible'} ${
-              size === 'large' ? 'top-12 w-36' : 'top-20 w-16'
-            } absolute mt-2 rounded-lg border border-grey bg-white`}
-          >
-            <ul className="list-inside list-none ">
-              {items.map((item, index) => (
-                <li
-                  key={index.toString()}
-                  title={item}
-                  onClick={() => onClickItem(item)}
-                  className={`${
-                    size === 'large' ? 'text-sm' : 'text-[10px]'
-                  } cursor-pointer select-none truncate rounded-lg p-2 hover:bg-main/50`}
-                >
-                  {item}
-                </li>
-              ))}
-            </ul>
-          </div>,
-          containerRef.current
-        )}
-    </div>
+      <div
+        className={`${isOpen ? 'visible' : 'hidden'} ${
+          size === 'large' ? 'top-12 w-36' : 'top-20 w-16'
+        } absolute z-50 mt-2 rounded-lg border border-grey bg-white`}
+      >
+        <ul className="list-inside list-none ">
+          {items.map((item, index) => (
+            <li
+              key={index.toString()}
+              title={item}
+              onClick={() => onClickItem(item)}
+              className={`${
+                size === 'large' ? 'text-sm' : 'text-[10px]'
+              } cursor-pointer select-none truncate rounded-lg p-2 hover:bg-main/50`}
+            >
+              {item}
+            </li>
+          ))}
+        </ul>
+      </div>
+    </>
   );
 };
